# project-2-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "0.0.0"
    },
    "paths": {
...
        "/api/admin/users/": {
            "get": {
                "tags": [
                    "Admin Users Management"
                ],
                "summary": "show all users to admin",
                "description": "show all users to admin",
                "operationId": "9278e4d9b2d84c8f019aad2c2de63fff",
                "parameters": [
                    {
                        "name": "is_admin",
                        "in": "query",
                        "description": "choose: **true|false**"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "status: **is_active|is_blocked**",
                        "schema": {
                            "example": "is_active"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "user email",
                        "schema": {
                            "example": "name@gmail.com"
                        }
                    },
                    {
                        "name": "cell_phone_number",
                        "in": "query",
                        "description": "user cell phone",
                        "schema": {
                            "example": "09000000000"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "user name",
                        "schema": {
                            "example": "name"
                        }
                    },
                    {
                        "name": "data_from",
                        "in": "query",
                        "description": "list of users start from",
                        "schema": {
                            "example": "2025-01-07T18:18:28"
                        }
                    },
                    {
                        "name": "data_to",
                        "in": "query",
                        "description": "list of users data end to",
                        "schema": {
                            "example": "2025-01-07T18:18:28"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page",
                        "schema": {
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Access token for the created user.",
                                            "type": "string"
                                        },
                                        "data": {
                                            "description": "Access token for the created user.",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Users Management"
                ],
                "summary": "create a new user",
                "description": "create a new user",
                "operationId": "e2b18774acbbaaaeabbea38fc36beaa7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "saeideh khanramaki"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "khanramaki@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Aa123456"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Aa123456"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "09110000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "the user does not have this permission.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Access token for the created user.",
                                            "type": "string"
                                        },
                                        "user": {
                                            "description": "User details.",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/users/{user_id}/updateStatus": {
            "patch": {
                "tags": [
                    "Admin Users Management"
                ],
                "summary": "update an existing User's status",
                "description": "update an existing User'hps status",
                "operationId": "180f24bc609a1dc681074e107bc93ba0",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "user_id",
                        "required": true,
                        "example": "2"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "is_blocked": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "the user does not have this permission.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "user updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}